Authentication support for mlflow sync (#51)
* feat: authentication support for mlflow sync

* chore: formatting

* chore: changelog

* chore: changelog add link

* fix: sync authentication flag

* refactor: move `health_check` to submodule top level

* feat: add health check

* chore: update error msg

* refactor: mlflow utils
